requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
PyPDF2
python-docx
httpx
python-dotenv
openai
sentence-transformers  # optional: only needed if you enable embedding ranking
pytest
pytest-asyncio
